Version 2 - Fix a bugs in the SHA256 routine.
Version 2 fixes bugs in the SHA256 routine.  Although the old code did not create security issues, it is deemed best if the correct routine is used.  Update of all implementations is required, and if  any public keys or shared keys are different now they will need to be re-distributes.  Sorry.